requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
Brotli @ file:///croot/brotli-split_1736182456865/work
bs4==0.0.2
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.32.2
docker-pycreds==0.4.0
executing==2.2.0
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///croot/gmpy2_1738085463648/work
grpcio==1.70.0
h11==0.14.0
huggingface-hub==0.28.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.37.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1737760107953/work
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kagglehub==0.3.11
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
munch==4.0.0
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp312-cp312-linux_x86_64.whl#sha256=e2374991344fa2241a2153ef3d550d3a1cd2d50cb1f1d51eddc82870abc88021
nvidia-ml-py==12.570.86
nvitop==1.4.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.6
prettytable==3.14.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.2.1
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
selenium==4.30.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
stringzilla==3.12.5
supervision==0.25.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.14
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.1.0
ttach==0.0.3
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
urllib3 @ file:///croot/urllib3_1737133630106/work
wandb==0.19.8
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
yapf==0.43.0
zipp==3.21.0
